+++ widgets/agents/src/props/AgentsPropType.ts
@@ -1,14 +1,3 @@
-export type Agents = {
-    id: string;
-    ip: string;
-    deployment: string;
-    node: string;
-    system: string;
-    version: string;
-    // eslint-disable-next-line camelcase
-    install_method: string;
-};
-
 export default PropTypes.arrayOf(
     PropTypes.shape({
         id: PropTypes.string,
