+++ widgets/agents/src/types.ts
@@ -1,0 +1,10 @@
+export type Agents = {
+    id: string;
+    ip: string;
+    deployment: string;
+    node: string;
+    system: string;
+    version: string;
+    // eslint-disable-next-line camelcase
+    install_method: string;
+};
